refactor rules
