+++ src/rules.ts
@@ -1,87 +1,31 @@
+import {Rule, RuleContext} from "./rules/rule";
+import {ParagraphRule} from './rules/paragraph';
 import {
-    BlockElement,
-    CodeBlock,
-    Emphasis,
-    HeaderBlock,
-    Horizontal,
-    HTMLBlock,
-    ImageLink,
-    InlineCode,
-    InlineElement,
-    InlinePlain,
-    LateXBlock,
-    Link,
-    LinkDefinition,
-    ListBlock,
-    ListElement,
-    MathBlock,
-    MaybeToken,
-    NewLine,
-    Paragraph,
-    Quotes
-} from "./token";
-import {StringStream} from './source';
+    CodeBlockRule,
+    EmphasisRule,
+    HeaderBlockRule,
+    HorizontalRule,
+    HTMLBlockRule,
+    HTMLTagsRegexps,
+    InlineCodeRule,
+    InlinePlainExceptSpecialMarksRule,
+    InlinePlainRule,
+    LinkDefinitionRule,
+    LinkOrImageRule,
+    ListBlockRule,
+    NewLineRule,
+    OpenTagRegExp,
+    QuotesRule,
+    RegExpWithTagName
+} from "./rules/std";
+import {InlineLatexCommandRule, InlineMathRule, LatexBlockRule, MathBlockRule} from "./rules/latex";
+import {GFMFencedCodeBlockRule} from "./rules/gfm";
+
+export {RuleContext, Rule};
+export {HTMLTagsRegexps};
 
 
-export class RegExpWithTagName extends RegExp {
-    protected gIndex: number;
-
-    constructor(r: RegExp, gIndex: number) {
-        super(r);
-        this.gIndex = gIndex;
-    }
-
-    // noinspection JSUnusedGlobalSymbols
-    getTagName(g: RegExpExecArray): string {
-        return g[this.gIndex];
-    }
-}
-
-export class OpenTagRegExp extends RegExpWithTagName {
-    protected gOpenIndex: number;
-
-    constructor(r: RegExp, gIndex: number, gOpenIndex: number) {
-        super(r, gIndex);
-        this.gOpenIndex = gOpenIndex;
-    }
-
-    isSingleton(g: RegExpExecArray): boolean {
-        return g[this.gOpenIndex] !== '/'
-    }
-}
-
-export interface HTMLTagsRegexps {
-    validTags: RegExp;
-    open_tag: OpenTagRegExp;
-    close_tag: RegExpWithTagName;
-    comment: RegExp;
-    others: RegExp[];
-}
-
-
-export interface RuleContext {
-
-    parseBlockElement(source: StringStream): BlockElement
-
-    parseInlineElement(source: StringStream): InlineElement
-
-    parseBlockElements(source: StringStream): BlockElement[]
-
-    parseInlineElements(source: StringStream): InlineElement[]
-}
-
-export interface Rule {
-    readonly name?: string
-    readonly description?: string
-
-    match: (s: StringStream, ctx: RuleContext) => MaybeToken
-}
-
-function forward(s: StringStream, capturing: RegExpExecArray) {
-    s.forward(capturing[0].length)
-}
-
-const validTags: HTMLTagsRegexps = function (): HTMLTagsRegexps {
+export const validTags: HTMLTagsRegexps = function (): HTMLTagsRegexps {
     let validTags: string = /^<open_tag>|<close_tag>|<comment>|<processing_instruction>|<declaration>|<cdata>/.source;
     let open_tag: string = /(?:<(<tag_name>)(?:\s*<attribute>)*\s*(\/?)>)/.source;
     let close_tag: string = /(?:<\/(<tag_name>)\s*>)/.source;
@@ -126,546 +70,6 @@ const validTags: HTMLTagsRegexps = function (): HTMLTagsRegexps {
     };
 }();
 
-export class NewLineRule implements Rule {
-    readonly name: string = "NewLine";
-    readonly description: string = "Standard Markdown Block Rule";
-
-    public readonly regex: RegExp = /^\n+/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new NewLine(capturing[0]);
-    };
-}
-
-export class ParagraphRule implements Rule {
-    readonly name: string = "Paragraph";
-    readonly description: string = "Standard Markdown Block Rule";
-
-    // public readonly regex: RegExp = /^(?:(?:[^$]|\$(?!\$))(?:\n|$)?)+/;
-
-    protected readonly skipLaTeXBlock: boolean;
-
-    constructor(skipLaTeXBlock: boolean) {
-        this.skipLaTeXBlock = skipLaTeXBlock;
-    }
-
-    match(s: StringStream, ctx: RuleContext): MaybeToken {
-        let lastChar: string = 'a', i = 0;
-        if (s.source[0] == '\n') {
-            return undefined;
-        }
-        if (this.skipLaTeXBlock) {
-            for (; i < s.source.length; i++) {
-                if (lastChar === s.source[i] && (lastChar === '$' || lastChar === '\n')) {
-                    i--;
-                    break;
-                }
-                if (lastChar === '\n' && s.source[i] === '\\') {
-                    if (i + 1 < s.source.length &&
-                        (('a' <= s.source[i + 1] && s.source[i + 1] <= 'z') || ('A' <= s.source[i + 1] && s.source[i + 1] <= 'Z'))
-                    ) {
-                        i--;
-                        break;
-                    }
-                } else if (lastChar === '\\' && s.source[i] !== '\n') {
-                    lastChar = 'a';
-                } else {
-                    lastChar = s.source[i];
-                }
-            }
-        } else {
-            for (; i < s.source.length; i++) {
-                if (lastChar === s.source[i] && '$\n'.includes(lastChar)) {
-                    i--;
-                    break;
-                }
-                if (lastChar === '\\' && s.source[i] !== '\n') {
-                    lastChar = 'a';
-                } else {
-                    lastChar = s.source[i];
-                }
-            }
-        }
-        if (!i) {
-            return undefined;
-        }
-        let capturing = s.source.slice(0, i);
-        s.forward(i);
-        return new Paragraph(ctx.parseInlineElements(new StringStream(capturing)));
-    };
-}
-
-
-/* *+- */
-
-/* 1.   */
-export class ListBlockRule implements Rule {
-    readonly name: string = "ListBlock";
-    readonly description: string = "Standard Markdown Block Rule";
-    public static readonly blankRegex = /^[\t\v\f ]*\n/;
-    public static readonly listBlockRegex = /^([^\n]*(?:\n|$)(?:(?=[^\n0-9*+-])[^\n]*(?:\n|$))*)/;
-    public static readonly replaceRegex = /^(?: {4}|\t)/gm;
-
-    match(s: StringStream, ctx: RuleContext): MaybeToken {
-        let ordered: boolean;
-        if ("*+-".includes(s.source[0])) {
-            ordered = false;
-        } else if ('0' <= s.source[0] && s.source[0] <= '9') {
-            ordered = true;
-        } else {
-            return undefined;
-        }
-        return ListBlockRule.matchBlock(new ListBlock(ordered), s, ctx);
-    };
-
-    private static matchBlock(l: ListBlock, s: StringStream, ctx: RuleContext): ListBlock | undefined {
-        let nextMarker: string | undefined;
-        nextMarker = l.lookAhead(s);
-        if (!nextMarker) {
-            return undefined;
-        }
-        let blockContent: string, marker: string | undefined, lastSeparated = false;
-        for (blockContent = '', marker = nextMarker, nextMarker = undefined;
-             marker !== undefined;
-             blockContent = '', marker = nextMarker, nextMarker = undefined) {
-            do {
-                let capturing = ListBlockRule.listBlockRegex.exec(s.source);
-                if (capturing === null) {
-                    throw new Error("match block failed");
-                }
-                forward(s, capturing);
-                blockContent += capturing[0];
-            } while (l.lookAhead0(s) && (nextMarker = l.lookAhead(s)) === undefined);
-            let element = new ListElement(marker, ctx.parseBlockElements(
-                new StringStream(blockContent.replace(ListBlockRule.replaceRegex, '')),
-            ));
-            if (!nextMarker) {
-                let capturing = ListBlockRule.blankRegex.exec(s.source);
-                if (capturing !== null) {
-                    forward(s, capturing);
-                    element.blankSeparated = true;
-                    lastSeparated = true;
-                    if (l.lookAhead0(s)) {
-                        nextMarker = l.lookAhead(s);
-                    }
-                } else {
-                    element.blankSeparated = lastSeparated;
-                    lastSeparated = false;
-                }
-            }
-            l.listElements.push(element);
-        }
-        return l;
-    }
-}
-
-
-export class QuotesRule implements Rule {
-    readonly name: string = "Quotes";
-    readonly description: string = "Standard Markdown Block Rule";
-
-    public readonly regex: RegExp = /^( *>[^\n]*(?:\n[^\n]+)*\n?)+/;
-
-    match(s: StringStream, ctx: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new Quotes(ctx.parseBlockElements(
-            new StringStream(capturing[0].replace(/^ *> ?/gm, ''))
-        ));
-    };
-}
-
-export class HorizontalRule implements Rule {
-    readonly name: string = "Horizontal";
-    readonly description: string = "Standard Markdown Block Rule";
-
-    public readonly regex: RegExp = /^(?:(?:\*[\r\t ]*){3,}|(?:-[\r\t ]*){3,})(?:\n|$)/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new Horizontal();
-    };
-}
-
-
-export class CodeBlockRule implements Rule {
-    readonly name: string = "CodeBlock";
-    readonly description: string = "Standard Markdown Block Rule";
-
-    public static readonly regex: RegExp = /^((?: {4}|\t)[^\n]+(\n|$))+/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = CodeBlockRule.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new CodeBlock(capturing[0].replace(/^(?: {4}|\t)/gm, ''));
-    };
-}
-
-export class HeaderBlockRule implements Rule {
-    readonly name: string = "HeaderBlock";
-    readonly description: string = "Standard Markdown Block Rule";
-
-    public readonly atxRegex: RegExp = /^(#{1,6}) ([^\n]*?)#*(?:\n|$)/;
-    public readonly setextRegex: RegExp = /^([^\n]+)\n=+(?:\n|$)/;
-
-    match(s: StringStream, ctx: RuleContext): MaybeToken {
-        return this.matchATX(s, ctx) || this.matchSetext(s, ctx);
-    };
-
-    matchATX(s: StringStream, ctx: RuleContext): MaybeToken {
-        let capturing = this.atxRegex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-        forward(s, capturing);
-        return new HeaderBlock(ctx.parseInlineElements(new StringStream(capturing[2])), capturing[1].length);
-    }
-
-    matchSetext(s: StringStream, ctx: RuleContext): MaybeToken {
-        let capturing = this.setextRegex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-        forward(s, capturing);
-        return new HeaderBlock(ctx.parseInlineElements(new StringStream(capturing[1])), 1);
-    }
-}
-
-export class LinkDefinitionRule implements Rule {
-    readonly name: string = "LinkDefinition";
-    readonly description: string = "Standard Markdown Block Rule";
-    public readonly regex: RegExp = /^ *\[([^\]]+)]: *<?([^\s>]+)>?(?: +["'(]([^\n]*)["')])? *(?:\n|$)/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-        forward(s, capturing);
-        return new LinkDefinition(capturing[1], capturing[2], capturing[3]);
-    };
-}
-
-export class HTMLBlockRule implements Rule {
-    readonly name: string = "HTMLBlock";
-    readonly description: string = "Standard Markdown Block Rule";
-    private validTags: HTMLTagsRegexps;
-
-    constructor(validTags: HTMLTagsRegexps) {
-        this.validTags = validTags;
-    }
-
-
-    // public readonly singleTonRegex: RegExp = /^/;
-    // public readonly stdRegex: RegExp = /^/;
-
-    match(s: StringStream, ctx: RuleContext): MaybeToken {
-        let ot: OpenTagRegExp = this.validTags.open_tag;
-        let capturing = ot.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        if (ot.isSingleton(capturing)) {
-            return new HTMLBlock(capturing[0]);
-        }
-
-        return undefined
-    };
-}
-
-function _braceMatch(s: StringStream, l: string, r: string): string {
-    if (s.source[0] == l) {
-        let c: number = 0;
-        for (let j = 0; j < s.source.length; j++) {
-            if (s.source[j] == l) {
-                c++;
-            } else if (s.source[j] == r) {
-                c--;
-                if (c === 0) {
-                    let res = s.source.slice(0, j + 1);
-                    s.forward(j + 1);
-                    return res;
-                }
-            }
-        }
-        let res = s.source;
-        s.forward(s.source.length);
-        return res;
-    }
-    return '';
-}
-
-export class InlineLatexCommandRule implements Rule {
-    readonly name: string = "InlineLatexCommand";
-    readonly description: string = "Latex Inline Rule";
-
-    public static readonly cmdNameRegex = /^\\([a-zA-Z_]\w*)/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = InlineLatexCommandRule.cmdNameRegex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new InlinePlain(capturing[0] + this.braceMatch(s));
-    }
-
-    braceMatch(s: StringStream) {
-        let res: string = '';
-        for (let i = 0; !s.eof; i = 0) {
-            for (; !s.eof && ' \n\t\v\f\r'.includes(s.source[i]); i++) {
-            }
-            if (!'[{'.includes(s.source[i])) {
-                return res;
-            }
-            if (i) {
-                res += s.source.slice(0, i);
-                s.forward(i);
-            }
-            res += _braceMatch(s, '{', '}');
-            res += _braceMatch(s, '[', ']');
-        }
-        return res;
-    }
-}
-
-export class LatexBlockRule implements Rule {
-    readonly name: string = "LatexBlock";
-    readonly description: string = "Latex Inline Rule";
-
-    public static readonly cmdNameRegex = /^\\([a-zA-Z_]\w*)/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = InlineLatexCommandRule.cmdNameRegex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new LateXBlock(capturing[0] + this.braceMatch(s));
-    }
-
-    braceMatch(s: StringStream) {
-        let res: string = '';
-        for (let i = 0; !s.eof; i = 0) {
-            for (; !s.eof && ' \t\v\f\r'.includes(s.source[i]); i++) {
-            }
-            if (s.source[i] == '\n') {
-                i++;
-                if (s.source[i] == '\n') {
-                    s.forward(i);
-                    return res;
-                }
-            }
-            if (!'[{'.includes(s.source[i])) {
-                return res;
-            }
-            if (i) {
-                res += s.source.slice(0, i);
-                s.forward(i);
-            }
-            res += _braceMatch(s, '[', ']');
-            res += _braceMatch(s, '{', '}');
-        }
-        return res;
-    }
-}
-
-export class InlinePlainExceptSpecialMarksRule implements Rule {
-    readonly name: string = "InlinePlainExceptSpecialMarks";
-    readonly description: string = "Standard Markdown Inline Rule";
-
-    public readonly regex: RegExp = /^(?:\\[<`_*\[$\\]|[^<`_*\[$\\])+/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new InlinePlain(capturing[0]);
-    };
-}
-
-export class InlinePlainRule implements Rule {
-    readonly name: string = "InlinePlain";
-    readonly description: string = "Standard Markdown Inline Rule";
-
-    public readonly regex: RegExp = /^(?:[<`_*\[$\\](?:\\[<`_*\[$\\]|[^<`_*\[$\\])*|(?:\\[<`_*\[$\\]|[^<`_*\[$\\])+)/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new InlinePlain(capturing[0]);
-    };
-}
-
-export class InlineMathRule implements Rule {
-    readonly name: string = "InlineMath";
-    readonly description: string = "Markdown Inline Rule";
-
-    public readonly regex: RegExp = /^\$((?:[^$]|\\\$)+)\$/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new MathBlock(capturing[1], true);
-    };
-}
-
-export class MathBlockRule implements Rule {
-    readonly name: string = "MathBlock";
-    readonly description: string = "Markdown Block Rule";
-
-    public readonly regex: RegExp = /^\$\$((?:[^$]|\\\$)+)\$\$/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new MathBlock(capturing[1], false);
-    };
-}
-
-
-export class LinkOrImageRule implements Rule {
-    readonly name: string = "Link";
-    readonly description: string = "Standard Markdown Inline Rule";
-
-    public readonly regex: RegExp = /^(!?)\[((?:\[[^\]]*]|[^\[\]]|](?=[^\[]*]))*)]\(\s*<?([\s\S]*?)>?(?:\s+['"]([\s\S]*?)['"])?\s*\)/;
-    public readonly refRegex: RegExp = /^(!?)\[((?:\[[^\]]*]|[^\[\]]|](?=[^\[]*]))*)]\s*\[([^\]]*)]/;
-
-    match(s: StringStream, ctx: RuleContext): MaybeToken {
-        return this.matchInline(s, ctx) || this.matchRef(s, ctx);
-    };
-
-    matchInline(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-        forward(s, capturing);
-        if (capturing[1] !== '') {
-
-            return new ImageLink(capturing[2], capturing[3], true, capturing[4]);
-        } else {
-            return new Link(capturing[2], capturing[3], true, capturing[4]);
-        }
-    }
-
-    matchRef(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.refRegex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-        forward(s, capturing);
-        if (capturing[1] !== '') {
-            return new ImageLink(capturing[2], capturing[3], false);
-        } else {
-            return new Link(capturing[2], capturing[3], false);
-        }
-    }
-}
-
-
-export class EmphasisRule implements Rule {
-    readonly name: string = "Emphasis";
-    readonly description: string = "Standard Markdown Inline Rule";
-
-    public readonly regex: RegExp = /^(?:(_{1,2})([^_]+?)(_{1,2})|(\*{1,2})([^*]+?)(\*{1,2}))/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        let l: string = capturing[1] || capturing[4], r: string = capturing[3] || capturing[6];
-        if (l !== r) {
-            if (l.length < r.length) {
-                s.forward(capturing[0].length - 1);
-                return new Emphasis(capturing[2] || capturing[5], l.length);
-            }
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new Emphasis(capturing[2] || capturing[5], l.length);
-    };
-}
-
-
-export class InlineCodeRule implements Rule {
-    readonly name: string = "InlineCode";
-    readonly description: string = "Standard Markdown Inline Rule";
-
-    public readonly regex: RegExp = /^(?:``([^`\n\r\u2028\u2029](?:`?[^`\n\r\u2028\u2029])*)``|`([^`\n\r\u2028\u2029]+?)`)/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = this.regex.exec(s.source);
-        if (capturing === null) {
-            return undefined;
-        }
-
-        forward(s, capturing);
-        return new InlineCode(capturing[1] || capturing[2]);
-    };
-}
-
-
-export class GFMFencedCodeBlockRule implements Rule {
-    readonly name: string = "GFMCodeBlock";
-    readonly description: string = "GFM Markdown Block Rule";
-
-    public static readonly backtickRegex: RegExp = /^(`{3,}) *([^`\s]+)?[^`\n]*(?:\n|$)([\s\S]*?)(?:\1`*|$)/;
-    public static readonly tildeRegex: RegExp = /^(~{3,}) *([^~\s]+)?.*(?:\n|$)([\s\S]*?)(?:\1~*|$)/;
-
-    match(s: StringStream, _: RuleContext): MaybeToken {
-        let capturing = GFMFencedCodeBlockRule.backtickRegex.exec(s.source);
-        if (capturing === null) {
-            capturing = GFMFencedCodeBlockRule.tildeRegex.exec(s.source);
-            if (capturing === null) {
-                return undefined;
-            }
-        }
-
-        forward(s, capturing);
-        return new CodeBlock(capturing[3], capturing[2]);
-    };
-}
 
 export const inlineRules: Rule[] = newInlineRules();
 export const blockRules: Rule[] = newBlockRules();
@@ -675,6 +79,14 @@ export interface CreateBlockRuleOptions {
     enableLaTeX?: boolean;
     enableGFMRules?: boolean;
     validTags?: HTMLTagsRegexps;
+}
+
+export interface CreateInlineRuleOptions {
+    enableLaTeX?: boolean;
+}
+
+export interface CreateRuleOptions extends CreateInlineRuleOptions, CreateBlockRuleOptions {
+
 }
 
 // noinspection JSUnusedGlobalSymbols
@@ -696,7 +108,7 @@ export function newBlockRules(
     ];
 
     let rules2: Rule[] = [
-        new ParagraphRule(opts?.enableLaTeX || false),
+        new ParagraphRule({skipLaTeXBlock: opts?.enableLaTeX || false, skipMathBlock: true}),
     ];
 
     // default enable
@@ -709,10 +121,6 @@ export function newBlockRules(
     }
 
     return [...rules0, ...rules1, ...rules2];
-}
-
-export interface CreateInlineRuleOptions {
-    enableLaTeX?: boolean;
 }
 
 // noinspection JSUnusedGlobalSymbols
@@ -738,16 +146,9 @@ export function newInlineRules(
     return [...rules0, ...rules1];
 }
 
-export interface CreateRuleOptions extends CreateInlineRuleOptions, CreateBlockRuleOptions {
-
-}
-
 export function newRules(opts?: CreateRuleOptions) {
     return {
         inlineRules: newInlineRules(opts),
         blockRules: newBlockRules(opts),
     }
 }
-
-
-export {validTags};
